Fix pauses in step stream.
Two techniques used:
1) Rework acceleration calculations so that steppers finish at same time.
2) Keep moving at same speed between stepper move instructions.

First one is the more important one. Second one bridges the small ~0.1ms gaps
caused by the block changes.

No gaps at all visible on logic analyzer.
Not so sure if acceleration is correct in every situation, seems to do a bit
faster accels in some cases.

+++ src/libs/StepTicker.h
@@ -24,14 +24,14 @@ class StepTicker{
 
         StepTicker();
         ~StepTicker();
-        void set_frequency( float frequency );
+        void set_frequency( uint32_t frequency );
         void signal_a_move_finished();
         void set_reset_delay( float seconds );
         int register_motor(StepperMotor* motor);
         void add_motor_to_active_list(StepperMotor* motor);
         void remove_motor_from_active_list(StepperMotor* motor);
         void set_acceleration_ticks_per_second(uint32_t acceleration_ticks_per_second);
-        float get_frequency() const { return frequency; }
+        uint32_t get_frequency() const { return frequency; }
         void unstep_tick();
         uint32_t get_tick_cnt() const { return tick_cnt; }
         uint32_t ticks_since(uint32_t last) const { return (tick_cnt>=last) ? tick_cnt-last : (UINT32_MAX-last) + tick_cnt + 1; }
@@ -49,7 +49,7 @@ class StepTicker{
         friend class StepperMotor;
 
     private:
-        float frequency;
+        uint32_t frequency;
         uint32_t period;
         volatile uint32_t tick_cnt;
         std::vector<std::function<void(void)>> acceleration_tick_handlers;
